MetasharkDrf/requirements.txt
amqp==5.1.1
asgiref==3.5.2
aspose-cells==22.10.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.1.3
django-restframework==0.0.1
djangorestframework==3.14.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
excel2json-3==0.1.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
JPype1==1.4.1
kombu==5.2.4
MarkupSafe==2.1.1
openpyxl==3.0.10
packaging==21.3
prompt-toolkit==3.0.31
pyparsing==3.0.9
pytz==2022.6
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
